backend/requirements.txt
# Django后端依赖
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
Pillow>=10.0
gunicorn>=21.0
